+++ components/questions/Question.tsx
@@ -11,22 +11,22 @@ import { motion } from "framer-motion"
 import Link from "next/link"
 import { useMemo, useState } from "react"
 import { ErrorBoundary } from "react-error-boundary"
-import { QuestionWithStandardIncludes } from "../prisma/additional"
+import { QuestionWithStandardIncludes } from "../../prisma/additional"
+import { FormattedDate } from "../ui/FormattedDate"
+import { Username } from "../ui/Username"
+import { QuestionDetails } from "./QuestionDetails"
 import { ResolveButton } from "./ResolveButton"
 import { SharePopover } from "./SharePopover"
 import { UpdateableLatestForecast } from "./UpdateableLatestForecast"
-import { QuestionDetails } from "./questions/QuestionDetails"
-import { FormattedDate } from "./ui/FormattedDate"
-import { Username } from "./ui/Username"
 
 import { ArrowTopRightOnSquareIcon } from "@heroicons/react/24/solid"
 import toast from "react-hot-toast"
-import { getDateYYYYMMDD } from "../lib/_utils_common"
-import { getQuestionUrl } from "../lib/web/question_url"
-import { api } from "../lib/web/trpc"
-import { invalidateQuestion, useUserId } from "../lib/web/utils"
-import { QuestionDetailsOption } from "./questions/QuestionDetailsOptions"
-import { InfoButton } from "./ui/InfoButton"
+import { getDateYYYYMMDD } from "../../lib/_utils_common"
+import { getQuestionUrl } from "../../lib/web/question_url"
+import { api } from "../../lib/web/trpc"
+import { invalidateQuestion, useUserId } from "../../lib/web/utils"
+import { InfoButton } from "../ui/InfoButton"
+import { QuestionDetailsOption } from "./QuestionDetailsOptions"
 
 export function Question({
   question,
