+++ components/questions/UpdateableLatestForecast.tsx
@@ -1,16 +1,16 @@
+import { Forecast, QuestionOption } from "@prisma/client"
+import { Decimal } from "@prisma/client/runtime/library"
 import clsx from "clsx"
 import { useRef, useState } from "react"
 import { useDebouncedCallback } from "use-debounce"
 import {
   displayForecast,
   getMostRecentForecastForUser,
-} from "../lib/_utils_common"
-import { sendToHost } from "../lib/web/embed"
-import { api } from "../lib/web/trpc"
-import { invalidateQuestion, useUserId } from "../lib/web/utils"
-import { QuestionWithStandardIncludes } from "../prisma/additional"
-import { Forecast, QuestionOption } from "@prisma/client"
-import { Decimal } from "@prisma/client/runtime/library"
+} from "../../lib/_utils_common"
+import { sendToHost } from "../../lib/web/embed"
+import { api } from "../../lib/web/trpc"
+import { invalidateQuestion, useUserId } from "../../lib/web/utils"
+import { QuestionWithStandardIncludes } from "../../prisma/additional"
 
 function closeLinkPopup() {
   sendToHost("close_link_popup")
